Scrape 'file -v' for default magic file.
This works back to 3.28, which is more than we need because libmagic
itself was not introduced until 4.0 (author date 2003-03-24). But we
have to delete $ENV{MAGIC} to get the answer we want, otherwise we may
just get the contents of that environment variable.

THe libmagic library contains magic_getpath() which can also be used to
access this information (again after deleting $ENV{MAGIC}). But this is
undocumentedm and was not introduced until 5.06 (author date
2011-04-14).