Drop content of generated PDF files
Leave only the header, which is needed for the tests.

The previous version of the file was generated and the LibreOffice source
which is not in version control.

The previous version of the file was also much larger, which is a waste of
space in the tarball.

See-also: https://wiki.debian.org/UpstreamGuide
See-also: https://github.com/houseabsolute/File-LibMagic/issues/17