Drop content of generated PDF file
Leave only the header, which is needed for the tests.

The previous version of the file was generated and the
LibreOffice source of the file is missing from the VCS.

The previous version of the file was also much larger,
which is a waste of space in the tarball.

See-also: https://wiki.debian.org/UpstreamGuide
See-also: https://github.com/houseabsolute/File-LibMagic/issues/17